requirements.txt
alembic==1.12.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.3
dnspython==2.6.1
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.2
fastapi-sessions==0.3.2
fastavro==1.9.4
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.3.0
pydantic_core==2.6.3
pydispatch==1.1.0
PyDispatcher==2.0.7
pytest==8.0.2
pytest-cov==4.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.20
sqlalchemy-pagination==0.0.2
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
pulsar-client[avro]==3.4.0

# functions runtime
pulsar-client[functions]==3.4.0

# all optional components
pulsar-client[all]==3.4.0